Add files via upload
Using requests and BeautifulSoap libraries,this web scrapping script has been written in python on Jupyter Notebook. Web scraping refers to the extraction of data from a website. This information is collected and then exported into a format that is more useful for the user. Be it a spreadsheet or an API. For learning purpose,code has been written in python and used external web url.Don't try to do it for any illegal activity.